html/semantics/forms/form-submission-0/form-submission-algorithm.html

<!DOCTYPE html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#form-submission-algorithm">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/targetted-form.js"></script>
<body>
<script>
test(() => {
  let form = populateForm('<input name=n10 value=v10>');
  let counter = 0;
  form.addEventListener('formdata', e => {
    ++counter;
    form.submit();
  });
  form.submit();
  assert_equals(counter, 1);
  new FormData(form);
  assert_equals(counter, 2);
}, 'If constructing entry list flag of form is true, then return');


async_test(t => {
  let form = populateForm('<input type=hidden name=n value=v0><input type=submit>');
  let submitter = form.querySelector('[type=submit]');
  let iframe = form.previousSibling;
  form.onsubmit = () => {
    let hidden = form.querySelector('[type=hidden]');
    hidden.value = 'inner click()';
    submitter.click();
    hidden.value = 'v2';
  };
  submitter.click();
  iframe.onload = t.step_func_done(() => {
    assert_not_equals(iframe.contentWindow.location.search.indexOf('n=v2'), -1);
  });
}, "If form's firing submit event is true, then return");

let test9 = async_test('Cannot navigate (after constructing the entry list)');
test9.step(() => {
  let form = populateForm('<input name=n1 value=v1>');
  form.onformdata = (e) => { e.target.remove(); };
  let wasLoaded = false;
  let iframe = form.previousSibling;
  // Request to load '/common/dummy.xhtml', and immediately submit the form to
  // the same frame. If the form submission is aborted, the first request
  // will be completed.
  iframe.onload = test9.step_func_done(() => {
    wasLoaded = true;
    assert_true(iframe.contentWindow.location.search.indexOf('n1=v1') == -1);
  });
  iframe.src = '/common/dummy.xhtml';
  assert_false(wasLoaded, 'Make sure the first loading is ongoing.');
  form.submit();
});
</script>
</body>
